==> packages/wmb/wmb-0.1.4.tar.gz/wmb-0.1.4/analysis/integration/mc_3c/Template/06.confusion_matrix.ipynb <==
from ALLCools.mcds import MCDS
from ALLCools.plot import *
from ALLCools.integration import calculate_overlap_score
from wmb import cemba, aibs, brain

import pathlib
import pandas as pd
import numpy as np
import anndata
import matplotlib.pyplot as plt
import seaborn as sns
import networkx as nx

dataset = 'CEMBA_3C'
ref_label = 'mC'
query_label = 'm3C'
categorical_key = ['L1_annot', 'L2', 'DissectionRegion']

## Load

mc_annot = cemba.get_mc_annot()
m3c_annot = cemba.get_m3c_annot()

adata_merge = anndata.read_h5ad('final_with_coords.h5ad')

adata_merge

mc_adata = adata_merge[adata_merge.obs['Modality'] == ref_label].copy()
m3c_adata = adata_merge[adata_merge.obs['Modality'] == query_label].copy()

mc_meta = adata_merge.obs[adata_merge.obs['Modality'] == ref_label].copy()
m3c_meta = adata_merge.obs[adata_merge.obs['Modality'] == query_label].copy()

## Scatter

fig, axes = plt.subplots(figsize=(8, 4 * len(categorical_key)),
                         dpi=300,
                         ncols=2,
                         nrows=len(categorical_key))

for ax_row, cat_key in zip(axes, categorical_key):
    ax = ax_row[0]
    if m3c_meta[cat_key].unique(
    ).size > 30 or m3c_meta[f'{cat_key}_transfer'].unique().size > 30:
        text_anno = False
    else:
        text_anno = True

    categorical_scatter(ax=ax,
                        data=m3c_adata,
                        coord_base='tsne',
                        palette='tab20',
                        text_anno=m3c_meta[cat_key] if text_anno else None,
                        hue=m3c_meta[cat_key],
                        max_points=None)
    ax.set(title=cat_key)

    ax = ax_row[1]
    categorical_scatter(
        ax=ax,
        data=m3c_adata,
        coord_base='tsne',
        palette='tab20',
        text_anno=m3c_meta[f'{cat_key}_transfer'] if text_anno else None,
        hue=m3c_meta[f'{cat_key}_transfer'],
        max_points=None)
    ax.set(title=f'{cat_key} Transfer')

## Confusion Matrix Based On Label Transfer

for cat_key in categorical_key:
    label_transfer = pd.read_hdf(f'{cat_key}_transfer.hdf')
    label_confusion = label_transfer.groupby(
        m3c_annot[cat_key].to_pandas().astype(str)).mean()
    label_confusion.to_hdf(f'{cat_key}.label_transfer.confusion.hdf', key='data')
    if label_confusion.shape[1] == 1 or label_confusion.shape[0] == 1:pass
    else:g = sns.clustermap(data=label_confusion, vmax=0.5)
    

## Confusion Matrix Based On Co-clustering Overlap Score

m3c_meta['cocluster'] = adata_merge.obs['leiden']
mc_meta['cocluster'] = adata_merge.obs['leiden']

for cat_key in categorical_key:
    os_mat = calculate_overlap_score(m3c_meta[[cat_key, 'leiden']],
                                     mc_meta[[cat_key, 'leiden']])
    os_mat.to_hdf(f'{cat_key}.overlap_score.hdf', key='data')
    if os_mat.shape[1] == 1 or os_mat.shape[0] == 1:pass
    else:g = sns.clustermap(data=os_mat, vmax=0.5)

!touch finish
